requirements.txt
flask
python-dotenv
requests
scapy
python-nmap
shodan
plotly
